+++ src/Core/Dfe.Complete.Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -32,6 +32,7 @@ namespace Dfe.Complete.Infrastructure
             //Queries
             services.AddScoped<IListAllProjectsQueryService, ListAllProjectsQueryService>();
             services.AddScoped<IConversionCsvQueryService, ConversionCsvQueryService>();
+            services.AddScoped<ILocalAuthoritiesQueryService, LocalAuthoritiesQueryService>();
 
             // Authentication
             //services.AddCustomAuthorization(config);