+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Customizations/CustomWebApplicationDbContextFactoryCustomization.cs
@@ -113,6 +113,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
                 services.AddCompleteApiClient<IProjectsClient, ProjectsClient>(config, client);
                 services.AddCompleteApiClient<ICsvExportClient, CsvExportClient>(config, client);
                 services.AddCompleteApiClient<IUsersClient, UsersClient>(config, client);
+                services.AddCompleteApiClient<IServiceSupportClient, ServiceSupportClient>(config, client);
                 var serviceProvider = services.BuildServiceProvider();
 
                 fixture.Inject(factory);
@@ -121,6 +122,7 @@ namespace Dfe.Complete.Api.Tests.Integration.Customizations
                 fixture.Inject(serviceProvider.GetRequiredService<IProjectsClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<ICsvExportClient>());
                 fixture.Inject(serviceProvider.GetRequiredService<IUsersClient>());
+                fixture.Inject(serviceProvider.GetRequiredService<IServiceSupportClient>());
                 fixture.Inject(new List<Claim>());
 
                 return factory;